Introduce ProjectLoader class to handle responsibilities of loading a project
