+++ src/OmniSharp.MSBuild/ProjectFile/ItemNames.cs
@@ -1,0 +1,11 @@
+﻿namespace OmniSharp.MSBuild.ProjectFile
+{
+    internal static class ItemNames
+    {
+        public const string Analyzer = nameof(Analyzer);
+        public const string Compile = nameof(Compile);
+        public const string PackageReference = nameof(PackageReference);
+        public const string ProjectReference = nameof(ProjectReference);
+        public const string ReferencePath = nameof(ReferencePath);
+    }
+}
