+++ src/OmniSharp.MSBuild/ProjectFile/MetadataNames.cs
@@ -1,0 +1,11 @@
+﻿namespace OmniSharp.MSBuild.ProjectFile
+{
+    internal static class MetadataNames
+    {
+        public const string FullPath = nameof(FullPath);
+        public const string IsImplicitlyDefined = nameof(IsImplicitlyDefined);
+        public const string Project = nameof(Project);
+        public const string ReferenceSourceTarget = nameof(ReferenceSourceTarget);
+        public const string Version = nameof(Version);
+    }
+}
