--- src/OmniSharp.MSBuild/ProjectFile/ProjectFileInfo.ItemNames.cs
@@ -1,14 +1,0 @@
-﻿namespace OmniSharp.MSBuild.ProjectFile
-{
-    public partial class ProjectFileInfo
-    {
-        private static class ItemNames
-        {
-            public const string Analyzer = nameof(Analyzer);
-            public const string Compile = nameof(Compile);
-            public const string PackageReference = nameof(PackageReference);
-            public const string ProjectReference = nameof(ProjectReference);
-            public const string ReferencePath = nameof(ReferencePath);
-        }
-    }
-}
