--- src/OmniSharp.MSBuild/ProjectFile/ProjectFileInfo.MetadataNames.cs
@@ -1,14 +1,0 @@
-﻿namespace OmniSharp.MSBuild.ProjectFile
-{
-    public partial class ProjectFileInfo
-    {
-        private static class MetadataNames
-        {
-            public const string FullPath = nameof(FullPath);
-            public const string IsImplicitlyDefined = nameof(IsImplicitlyDefined);
-            public const string Project = nameof(Project);
-            public const string ReferenceSourceTarget = nameof(ReferenceSourceTarget);
-            public const string Version = nameof(Version);
-        }
-    }
-}
