--- src/OmniSharp.MSBuild/ProjectFile/ProjectFileInfo.PropertyNames.cs
@@ -1,43 +1,0 @@
-﻿namespace OmniSharp.MSBuild.ProjectFile
-{
-    public partial class ProjectFileInfo
-    {
-        private static class PropertyNames
-        {
-            public const string AllowUnsafeBlocks = nameof(AllowUnsafeBlocks);
-            public const string AssemblyName = nameof(AssemblyName);
-            public const string AssemblyOriginatorKeyFile = nameof(AssemblyOriginatorKeyFile);
-            public const string BuildProjectReferences = nameof(BuildProjectReferences);
-            public const string BuildingInsideVisualStudio = nameof(BuildingInsideVisualStudio);
-            public const string Configuration = nameof(Configuration);
-            public const string CscToolExe = nameof(CscToolExe);
-            public const string CscToolPath = nameof(CscToolPath);
-            public const string DefineConstants = nameof(DefineConstants);
-            public const string DesignTimeBuild = nameof(DesignTimeBuild);
-            public const string DocumentationFile = nameof(DocumentationFile);
-            public const string LangVersion = nameof(LangVersion);
-            public const string OutputType = nameof(OutputType);
-            public const string MSBuildExtensionsPath = nameof(MSBuildExtensionsPath);
-            public const string MSBuildSDKsPath = nameof(MSBuildSDKsPath);
-            public const string NoWarn = nameof(NoWarn);
-            public const string OutputPath = nameof(OutputPath);
-            public const string Platform = nameof(Platform);
-            public const string ProjectAssetsFile = nameof(ProjectAssetsFile);
-            public const string ProvideCommandLineArgs = nameof(ProvideCommandLineArgs);
-            public const string ProjectGuid = nameof(ProjectGuid);
-            public const string ProjectName = nameof(ProjectName);
-            public const string _ResolveReferenceDependencies = nameof(_ResolveReferenceDependencies);
-            public const string RoslynTargetsPath = nameof(RoslynTargetsPath);
-            public const string SignAssembly = nameof(SignAssembly);
-            public const string SkipCompilerExecution = nameof(SkipCompilerExecution);
-            public const string SolutionDir = nameof(SolutionDir);
-            public const string TargetFramework = nameof(TargetFramework);
-            public const string TargetFrameworkMoniker = nameof(TargetFrameworkMoniker);
-            public const string TargetFrameworkRootPath = nameof(TargetFrameworkRootPath);
-            public const string TargetFrameworks = nameof(TargetFrameworks);
-            public const string TargetPath = nameof(TargetPath);
-            public const string VisualStudioVersion = nameof(VisualStudioVersion);
-            public const string VsInstallRoot = nameof(VsInstallRoot);
-        }
-    }
-}
