--- src/OmniSharp.MSBuild/ProjectFile/ProjectFileInfo.TargetNames.cs
@@ -1,12 +1,0 @@
-﻿namespace OmniSharp.MSBuild.ProjectFile
-{
-    public partial class ProjectFileInfo
-    {
-        private static class TargetNames
-        {
-            public const string Compile = nameof(Compile);
-            public const string CoreCompile = nameof(CoreCompile);
-            public const string ResolveReferences = nameof(ResolveReferences);
-        }
-    }
-}
