+++ src/OmniSharp.MSBuild/ProjectFile/TargetNames.cs
@@ -1,0 +1,9 @@
+﻿namespace OmniSharp.MSBuild.ProjectFile
+{
+    internal static class TargetNames
+    {
+        public const string Compile = nameof(Compile);
+        public const string CoreCompile = nameof(CoreCompile);
+        public const string ResolveReferences = nameof(ResolveReferences);
+    }
+}
